Comment validation with JavaScript
Validate comment form with JavaScript. It was quite confusing to not
get any feedback if the comment was invalid. Now alert box will be
raised if some of the comment fields are invalid. JavaScript is in the
Python source code which is not the best way, but is sufficient for
now.